🛠 Extract tile generation and coordinate translation from move
A lot of move's code had little to do with actually moving from a to b,
and had more to do with generating tiles and doing the math to translate
coordinates.

By splitting these out into their own files; I hope to make the behavior
of `move.js` a bit clearer.