Hack out an initial grid presentation layer
I figured it'd be a little easier to figure out WTF is going on if I had
a visualization of the resulting grid working.

I exported the SVG from Concepts, imported it into Vectornator,
and one-by-one went through and made sure they were 150px wide with an
outlined border (rather than stroked).

Next step will be to either:

- Move more of the tiles through the pipeline
- Start implmenting the "Does it fit" algorithm for filtering which
  tiles can go where